+++ gems/aws-sdk-paymentcryptography/CHANGELOG.md
@@ -1,6 +1,11 @@
 Unreleased Changes
 ------------------
 
+1.53.0 (2026-04-03)
+------------------
+
+* Feature - Adds optional support to retrieve previously generated import and export tokens to simplify import and export functions
+
 1.52.0 (2026-03-18)
 ------------------